python-api/requirements.txt
Flask
certifi==2025.6.15
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
fonttools==4.57.0
idna==3.10
joblib==1.4.2
kiwisolver==1.4.8
lightgbm==4.6.0
matplotlib==3.10.1
numpy==2.2.4
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psycopg2-binary==2.9.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.5.0
xgboost==3.0.0
weasyprint
